+++ test/python/loading_tests.py
@@ -42,6 +42,7 @@ class LoadingTestCaseSuite(unittest.TestCase):
         cfg.quote_char = '"'
         cfg.data_loading_error = False
         cfg.debug = False
+        cfg.truncate_before_load = False
 
     def test_loading_MySQL(self):
         print("test_loading_MySQL")
@@ -247,6 +248,68 @@ class LoadingTestCaseSuite(unittest.TestCase):
                               "-s", "	"]
                              )
                          )
+
+    def test_truncate_table_before_load_Oracle(self):
+        print("test_truncate_table_before_load_Oracle")
+        truncate_table_before_load(self, f.DBType.ORACLE.value)
+
+    def test_truncate_table_before_load_MySQL(self):
+        print("test_truncate_table_before_load_MySQL")
+        truncate_table_before_load(self, f.DBType.MYSQL.value)
+
+    def test_truncate_table_before_load_Postgres(self):
+        print("test_truncate_table_before_load_Postgres")
+        truncate_table_before_load(self, f.DBType.POSTGRES.value)
+
+    def test_truncate_table_before_load_SqlServer(self):
+        print("test_truncate_table_before_load_SQLServer")
+        truncate_table_before_load(self, f.DBType.SQLSERVER.value)
+
+    def test_truncate_table_before_load_Db2(self):
+        print("test_truncate_table_before_load_Db2")
+        truncate_table_before_load(self, f.DBType.DB2.value, login["db2_user"])
+
+
+def truncate_table_before_load(self, db_type, username=login["user"]):
+
+    params = ["load",
+              "-f", "../resources/201811-citibike-tripdata.csv*",
+              "-o", db_type,
+              "-u", username,
+              "-p", login["password"],
+              "-d", login["database"],
+              "-t", login["table"],
+              "--truncate"
+              ]
+
+    self.assertEqual(f.ExitCodes.SUCCESS.value, csv2db.run(params))
+    # MySQL: a connection cannot hold the talbe still in cache
+    # otherwise the TRUNCATE TABLE will hang (as it is a DROP/CREATE TABLE)
+    count1 = get_count_from_table(db_type, username, login["table"])
+
+    # Reset global truncate parameter
+    cfg.truncate_before_load = False
+
+    self.assertEqual(f.ExitCodes.SUCCESS.value, csv2db.run(params))
+    count2 = get_count_from_table(db_type, username, login["table"])
+
+    self.assertEqual(count1, count2)
+
+
+def get_count_from_table(db_type, username, table):
+    conn = f.get_db_connection(db_type,
+                               username,
+                               login["password"],
+                               "localhost",
+                               f.get_default_db_port(db_type),
+                               login["database"])
+    cur = conn.cursor()
+    cur.execute("SELECT COUNT(1) FROM {0}".format(login["table"]))
+    count = cur.fetchone()[0]
+    cur.close()
+    conn.close()
+
+    return count
 
 
 if __name__ == '__main__':